requirements.txt
torch>=1.13.1
huggingface-hub>=0.24.7,<0.25
transformers>=4.43.0,<=4.45.0
datasets>=2.14.3
accelerate>=0.27.2
loguru==0.7.0
tqdm==4.62.3
click
sentencepiece
protobuf
tiktoken
einops
transformers_stream_generator
tenacity
python-dotenv
peft>=0.10.0
gitpython
pre-commit
modal